sound-row: Swith to GtkFileDialog
This makes us use the portal and hence we get an adaptive file dialog in
most distros. The downside is that preselecting the old file doesn't
work anymore atm.

We bump the required GTK version for 4.12 to that. We require 4.12
implicitly anyway due to libadwaita 1.4.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/87>