application: Don't use self for GApplication
self refers to an object of type PhoshMobileSettingsApplication not
GApplication. Adjust the variable names in functions that did this
differently. This makes it obvious where casts are needed.

While at that fix an overly long line by introducing a variable.

Fixes: 6f40214 ("application: Override handle_local_options vfunc") and
       related commits.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/109>