lockscreen-panel: Only show lockscreen plugins
With more plugin types coming to phosh via

https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1359

we want to make sure we only list the plugins with a "lockscreen"
type here.

For a transition phase we assume that no type means "lockscreen".  This
can be dropped in the future.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/113>