panel-switcher: Allow going back to same page in folded mode
Rename on_visible_child_changed() in MobileSettingsWindow to
show_content_cb(), since that's what it is doing.

Connect this signal handler also to the 'row-activated' signal of
MsPanelSwitcher so that the content page is shown in folded mode also
when the visible child of the stack does not change.

Closes: https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/issues/38
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/139>