osk-panel: Better handle the default completer
The default completer is indicated via the empty string (''). Make sure
we provide some meaningful information.

We do that here (rather than via the OSK) as for the case of a user
selection we can then drop the confusing default entry.

At some point we will enforce minimal completers on the OSK side and
then this entry will just not be shown anymore.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/163>