panel-switcher: Make scrollable
This allows the switcher to be scolled if the height of the program is
too small to display all options, which can happen if a device like the
Librem 5 is viewing the program in landscape mode. It might also help
devices with shorter screen if users have chosen a large text font.
Additionally, it faciliates adding more options to the list in the
future.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/167>