doap: Fix name, shortname, shortdesc and a typo
This project is not called Phosh, so let's set it to the real name.
Makes for a slightly less confusing experience when selecting project in
GNOME Builder.

Additionally, make the shortdesc something unique.

And, finally, fix a typo in descrption (up -> app).

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh-mobile-settings/-/merge_requests/168>